add check_poison function
Probably contains a whole host of bugs at this point.

Had to add/change a lot of stuff for this:
1. `IPMultiple` now has `add()` and `remove()` methods
2. `IPAddress.poisoned` is now a dict, `IPAddress.poisoned_arp`
containing all fake arp entries on the host having addr IPAddress.

Onto the actual function:
The source for this function is heavily documented, if only for my own
sanity. I'm gonna cite DRY and not do document it here again.